ARROW-16158: Had to move engine/visibility to engine/substrait/visibility since engine is not published yet.  In the future we can move it back as visibility.h is not very public and no one should be using it anyways.
