+++ cpp/src/arrow/engine/substrait/extension_set.h
@@ -21,7 +21,7 @@
 
 #include <vector>
 
-#include "arrow/engine/visibility.h"
+#include "arrow/engine/substrait/visibility.h"
 #include "arrow/type_fwd.h"
 #include "arrow/util/optional.h"
 #include "arrow/util/string_view.h"
