+++ cpp/src/arrow/engine/substrait/plan_internal.h
@@ -20,7 +20,7 @@
 #pragma once
 
 #include "arrow/engine/substrait/extension_set.h"
-#include "arrow/engine/visibility.h"
+#include "arrow/engine/substrait/visibility.h"
 #include "arrow/type_fwd.h"
 
 #include "substrait/plan.pb.h"  // IWYU pragma: export
