+++ cpp/src/arrow/engine/substrait/type_internal.h
@@ -22,7 +22,7 @@
 #include <utility>
 
 #include "arrow/engine/substrait/extension_set.h"
-#include "arrow/engine/visibility.h"
+#include "arrow/engine/substrait/visibility.h"
 #include "arrow/type_fwd.h"
 
 #include "substrait/type.pb.h"  // IWYU pragma: export
