+++ cpp/src/arrow/engine/substrait/visibility.h
@@ -15,6 +15,8 @@
 // specific language governing permissions and limitations
 // under the License.
 
+// TODO(westonpace): Once we have a propert engine module this file
+// should be renamed arrow/engine/visibility.h
 // This API is EXPERIMENTAL.
 
 #pragma once
